projects/builds/pyproject.toml
[project]
name = "builds"
version = "0.1.0"
requires-python = ">=3.13"
authors = [
    { name="Tony Truong", email="serinth@gmail.com" }
]
dependencies = [
    "dynaconf>=3.2.6",
    "google-cloud-bigquery>=3.27.0",
    "requests>=2.32.3",
    "bigquery",
    "state-manager"
]

[tool.uv.sources]
bigquery = { workspace = true }
state-manager = { workspace = true }

[tool.uv.workspace]
members = ["../../libs/*"]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
]
